src replace part ipc into two platform, mv some function to impl_net
