+++ clashctl/backend/src/utils/backend.rs
@@ -1,16 +1,14 @@
-use std::{
-    io::Error,
-    path::{Path, PathBuf},
-};
+use std::path::{Path, PathBuf};
 mod impl_app;
 mod impl_clashsrv;
+mod impl_net;
 mod impl_profile;
 
 use super::{
     config::{CfgError, Config, ErrKind},
     parse_yaml,
 };
-use api::{ClashConfig, ClashUtil, Resp};
+use api::ClashUtil;
 
 const BASIC_FILE: &str = "basic_clash_config.yaml";
 
@@ -45,32 +43,6 @@ impl ClashBackend {
             },
             err_track,
         )
-    }
-    pub fn clash_version(&self) -> String {
-        match self.clash_api.version() {
-            Ok(v) => v,
-            Err(e) => {
-                log::warn!("{}", e);
-                "Unknown".to_string()
-            }
-        }
-    }
-    fn fetch_remote(&self) -> Result<ClashConfig, Error> {
-        use core::str::FromStr as _;
-        self.clash_api.config_get().and_then(|cur_remote| {
-            ClashConfig::from_str(cur_remote.as_str())
-                .map_err(|_| Error::new(std::io::ErrorKind::InvalidData, "Failed to prase str"))
-        })
-    }
-    pub fn restart_clash(&self) -> Result<String, Error> {
-        self.clash_api.restart(None)
-    }
-    fn dl_remote_profile(&self, url: &str) -> Result<Resp, Error> {
-        self.clash_api
-            .mock_clash_core(url, self.clash_api.version().is_ok())
-    }
-    fn config_reload(&self, body: String) -> Result<(), Error> {
-        self.clash_api.config_reload(body)
     }
     fn _update_state(
         &self,
